Delete EventFunnel API and introduce ThirdPartyContext
