+++ server/src/main/java/io/spine/server/BoundedContext.java
@@ -39,7 +39,6 @@ import io.spine.server.event.DelegatingEventDispatcher;
 import io.spine.server.event.EventBus;
 import io.spine.server.event.EventDispatcher;
 import io.spine.server.event.EventDispatcherDelegate;
-import io.spine.server.event.funnel.PostEvents;
 import io.spine.server.event.store.DefaultEventStore;
 import io.spine.server.integration.ExternalDispatcherFactory;
 import io.spine.server.integration.ExternalMessageDispatcher;
@@ -455,20 +454,6 @@ public abstract class BoundedContext implements AutoCloseable, Logging {
      */
     public boolean isMultitenant() {
         return spec.isMultitenant();
-    }
-
-    /**
-     * Obtains an instance of a fluent builder for manually posting events into this Context.
-     *
-     * <p>An event may be posted by hand:
-     * <ol>
-     *     <li>to an Aggregate, which {@linkplain io.spine.server.aggregate.Apply#allowImport
-     *         allows import} for this type of events;
-     *     <li>to external event reactors/subscribers.
-     * </ol>
-     */
-    public PostEvents postEvents() {
-        return new PostEvents(this);
     }
 
     /**
