--- server/src/test/java/io/spine/server/event/funnel/given/EditHistoryProjection.java
@@ -1,51 +1,0 @@
-/*
- * Copyright 2019, TeamDev. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.event.funnel.given;
-
-import io.spine.core.Subscribe;
-import io.spine.server.event.funnel.DocumentId;
-import io.spine.server.event.funnel.Edit;
-import io.spine.server.event.funnel.EditHistory;
-import io.spine.server.event.funnel.TextEdited;
-import io.spine.server.event.funnel.UserDeleted;
-import io.spine.server.projection.Projection;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class EditHistoryProjection
-        extends Projection<DocumentId, EditHistory, EditHistory.Builder> {
-
-    @Subscribe
-    void on(TextEdited event) {
-        builder()
-                .addEdit(event.getEdit());
-    }
-
-    @Subscribe(external = true)
-    void on(UserDeleted event) {
-        List<Edit> list = new ArrayList<>(builder().getEditList());
-        list.removeIf(edit -> edit.getEditor().equals(event.getUser()));
-        builder()
-                .clearEdit()
-                .addAllEdit(list);
-    }
-}
