+++ server/src/test/proto/spine/test/integration/doc_commands.proto
@@ -1,15 +1,15 @@
 syntax = "proto3";
 
-package spine.test.server.event.funnel;
+package spine.test.server.integration;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.server.event.funnel";
-option java_outer_classname = "CommandsProto";
+option java_package = "io.spine.server.integration";
+option java_outer_classname = "DocCommandsProto";
 option java_multiple_files = true;
 
-import "spine/test/event/funnel/docs.proto";
+import "spine/test/integration/docs.proto";
 
 message CreateDocument {
